refactor(PreOrderCursorIterator): use generator delegation
Refs #691
