+++ packages/apidom-parser-adapter-json/test/syntactic-analysis/PreOrderCursorChildrenIterator.ts
@@ -3,7 +3,6 @@ import { sexprs } from '@swagger-api/apidom-core';
 
 import { lexicalAnalysis, syntacticAnalysisDirect } from '../../src/adapter-node';
 import PreOrderCursorChildrenIterator from '../../src/syntactic-analysis/PreOrderCursorChildrenIterator';
-import PreOrderCursorIterator from '../../src/syntactic-analysis/PreOrderCusrorIterator';
 
 describe('syntactic-analysis', function () {
   context('PreOrderCursorChildrenIterator', function () {
@@ -15,17 +14,6 @@ describe('syntactic-analysis', function () {
       const apiDOM = syntacticAnalysisDirect(optimizedCst);
 
       expect(sexprs(apiDOM)).toMatchSnapshot();
-    });
-  });
-
-  context('PreOrderCursorIterator', function () {
-    specify('should create optimized list of surrogate syntax nodes', async function () {
-      const cst = await lexicalAnalysis('[1, 2]');
-      const cursor = cst.walk();
-      const iterator = new PreOrderCursorIterator(cursor);
-      const optimizedList = [...iterator];
-
-      expect(optimizedList).to.have.lengthOf(7);
     });
   });
 });
